Move the "golden" package below "testutils"
