+++ internal/services/manager_test.go
@@ -12,10 +12,10 @@ import (
 
 	"github.com/stretchr/testify/require"
 	"github.com/ubuntu/authd"
-	"github.com/ubuntu/authd/internal/golden"
 	"github.com/ubuntu/authd/internal/services"
 	"github.com/ubuntu/authd/internal/services/errmessages"
 	"github.com/ubuntu/authd/internal/testutils"
+	"github.com/ubuntu/authd/internal/testutils/golden"
 	"github.com/ubuntu/authd/internal/users"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
