+++ internal/users/localgroups/localgroups_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
-	"github.com/ubuntu/authd/internal/golden"
+	"github.com/ubuntu/authd/internal/testutils/golden"
 	"github.com/ubuntu/authd/internal/users/localgroups"
 	localgroupstestutils "github.com/ubuntu/authd/internal/users/localgroups/testutils"
 )
